added enemy animation
