+++ pages/index.js
@@ -12,7 +12,7 @@ import ShowDamage from "@/components/ShowDamage";
 import Pouch from "@/components/Pouch";
 import AudioHandler from "@/components/AudioHandler";
 import styled from "styled-components";
-import { Howl, Howler } from "howler";
+import { fightSound, hurtSound } from "@/components/soundHandler";
 
 const EnemyBox = styled.div`
   display: flex;
@@ -31,7 +31,6 @@ const BackgroundAndEnemy = styled.div`
 
 const LevelBox = styled.div`
   display: flex;
-
   top: 0;
 `;
 
@@ -75,43 +74,6 @@ export default function HomePage({ library, userData, itemList, currentBook }) {
   const [inventorySlots, setInventorySlots] = useState(
     userData[0].books[0].inventorySlots
   );
-
-  function fightSound() {
-    const fightSounds = [
-      "/assets/audio/fx/fight01.mp3",
-      "/assets/audio/fx/fight02.mp3",
-      "/assets/audio/fx/fight03.mp3",
-      "/assets/audio/fx/fight04.mp3",
-      "/assets/audio/fx/fight05.mp3",
-      "/assets/audio/fx/fight06.mp3",
-      "/assets/audio/fx/fight07.mp3",
-      "/assets/audio/fx/fight08.mp3",
-    ];
-
-    const randomFightSound = new Howl({
-      src: [fightSounds[Math.floor(Math.random() * fightSounds.length)]],
-    });
-
-    const randomVolume = Math.random() * 0.3;
-    randomFightSound.volume(randomVolume);
-
-    randomFightSound.play();
-  }
-  function hurtSound() {
-    const hurtSounds = [
-      "/assets/audio/fx/hurt01.mp3",
-      "/assets/audio/fx/hurt02.mp3",
-      "/assets/audio/fx/hurt03.mp3",
-    ];
-    const randomHurtSound = new Howl({
-      src: [hurtSounds[Math.floor(Math.random() * hurtSounds.length)]],
-    });
-
-    const randomVolume = Math.random() * 0.3;
-    randomHurtSound.volume(randomVolume);
-
-    randomHurtSound.play();
-  }
 
   useEffect(() => {
     setEnemyHealth(currentEnemy.health);
